some progress in splitting the shared component
