+++ src/counter/mod.rs
@@ -8,45 +8,15 @@ use bitcoin::sighash::{Prevouts, SighashCache};
 use bitcoin::taproot::{LeafVersion, TaprootBuilder};
 use bitcoin::transaction::Version;
 use bitcoin::{
-    Amount, OutPoint, ScriptBuf, Sequence, TapLeafHash, TapSighashType, Transaction, TxIn, TxOut,
-    Txid, Witness, WitnessProgram,
+    Amount, OutPoint, ScriptBuf, Sequence, TapLeafHash, TapSighashType, Transaction, TxIn, TxOut
+    , Witness, WitnessProgram,
 };
 use bitcoin_scriptexec::utils::scriptint_vec;
 use bitcoin_scriptexec::TxTemplate;
-use covenants_gadgets::internal_structures::cpp_int_32::CppInt32Gadget;
-use covenants_gadgets::structures::tagged_hash::{get_hashed_tag, HashTag, TaggedHashGadget};
-use covenants_gadgets::utils::pseudo::{OP_CAT2, OP_CAT3, OP_CAT4, OP_HINT};
-use covenants_gadgets::wizards::{tap_csv_preimage, tx};
+use covenants_gadgets::structures::tagged_hash::get_hashed_tag;
 use sha2::Digest;
 use std::str::FromStr;
-
-const DUST_AMOUNT: u64 = 330;
-
-/// Information necessary to create the new transaction.
-pub struct CovenantHints {
-    /// The counter value stored in the previous caboose.
-    pub prev_counter: u32,
-    /// The randomizer used in the previous caboose (for the Schnorr trick to work).
-    pub prev_randomizer: u32,
-    /// The balance of the previous program.
-    pub prev_balance: u64,
-    /// The txid of the previous program.
-    pub prev_txid: Txid,
-
-    /// The first input's outpoint of the transaction with txid.
-    pub prev_tx_outpoint1: OutPoint,
-    /// The second input's outpoint of the transaction with txid.
-    /// Note: the second input is optional.
-    pub prev_tx_outpoint2: Option<OutPoint>,
-
-    /// The second input in the new transaction, used to deposit more money into the program.
-    /// Note: The witness must be provided for this input.
-    pub optional_deposit_input: Option<TxIn>,
-
-    /// The balance of the new program, which needs to be smaller than the old balance plus the deposit,
-    /// but does not need to equal (some sats will be used to cover the transaction fee).
-    pub new_balance: u64,
-}
+use crate::common::{CovenantHints, covenants, DUST_AMOUNT};
 
 /// Compute the taproot of the script (which only has the script path, but not the key path).
 pub fn get_script_pub_key_and_control_block() -> (ScriptBuf, Vec<u8>) {
@@ -285,238 +255,8 @@ pub fn get_tx(info: &CovenantHints) -> (TxTemplate, u32) {
 }
 
 pub fn get_script() -> Script {
-    // Obtain the secp256k1 dummy generator, which would be point R in the signature, as well as
-    // the public key.
-    let secp256k1_generator = SECP256K1_GENERATOR.clone();
-
     script! {
-        // For more information about the construction of the Tap CheckSigVerify Preimage, please
-        // check out the `covenants-gadgets` repository.
-
-        { tap_csv_preimage::Step1EpochGadget::default() }
-        { tap_csv_preimage::Step2HashTypeGadget::from_constant(&TapSighashType::AllPlusAnyoneCanPay) }
-        { tap_csv_preimage::Step3VersionGadget::from_constant(&Version::ONE) }
-        { tap_csv_preimage::Step4LockTimeGadget::from_constant_absolute(&LockTime::ZERO) }
-        OP_CAT4
-
-        // current stack body: |1-4|
-
-        // first output: the same script itself
-
-        // get a hint: new balance (8 bytes)
-        OP_HINT
-        OP_SIZE 8 OP_EQUALVERIFY
-
-        // get a hint: this script's scriptpubkey (34 bytes)
-        OP_HINT
-        OP_SIZE 34 OP_EQUALVERIFY
-
-        // save a copy to the altstack
-        OP_DUP OP_TOALTSTACK
-
-        OP_PUSHBYTES_1 OP_PUSHBYTES_34
-        OP_SWAP OP_CAT3
-
-        OP_FROMALTSTACK OP_SWAP
-
-        // current stack body: |1-4|, this scriptpubkey, |output1|
-
-        // second output: the data carrier
-
-        // the balance must be DUST_AMOUNT
-        OP_PUSHBYTES_8 OP_PUSHBYTES_74 OP_PUSHBYTES_1 OP_PUSHBYTES_0 OP_PUSHBYTES_0
-        OP_PUSHBYTES_0 OP_PUSHBYTES_0 OP_PUSHBYTES_0 OP_PUSHBYTES_0
-        OP_CAT
-
-        // push the script hash header
-        OP_PUSHBYTES_2 OP_RETURN OP_PUSHBYTES_8
-
-        // get a hint: the new counter value in Bitcoin integer format (<=4 bytes)
-        OP_HINT
-        OP_1ADD OP_1SUB
-        OP_DUP 0 OP_GREATERTHAN OP_VERIFY
-
-        // save the new counter to the altstack
-        OP_DUP OP_TOALTSTACK
-
-        // get a hint: the old counter value in Bitcoin integer format (<=4 bytes)
-        OP_HINT
-        OP_1ADD OP_1SUB
-        OP_DUP 0 OP_GREATERTHANOREQUAL OP_VERIFY
-
-        // save the previous number into the altstack for later use
-        OP_DUP OP_TOALTSTACK
-        OP_TOALTSTACK
-
-        // extend the new counter to 4 bytes
-        { CppInt32Gadget::from_positive_bitcoin_integer() }
-
-        // get a hint: the randomizer for this transaction (4 bytes)
-        OP_HINT
-        OP_SIZE 4 OP_EQUALVERIFY
-        OP_CAT3
-
-        OP_SHA256
-
-        OP_PUSHBYTES_3 OP_PUSHBYTES_34 OP_PUSHBYTES_0 OP_PUSHBYTES_32
-        OP_SWAP OP_CAT3
-
-        OP_SHA256
-        OP_ROT OP_SWAP OP_CAT2
-
-        OP_FROMALTSTACK OP_SWAP
-
-        { tap_csv_preimage::Step7SpendTypeGadget::from_constant(1, false) } OP_CAT2
-
-        // current stack body: this scriptpubkey, previous counter, |1-7|
-
-        // get a hint: previous tx's txid
-        OP_HINT
-        OP_SIZE 32 OP_EQUALVERIFY
-
-        // save a copy to altstack
-        OP_DUP OP_TOALTSTACK
-
-        // require the output index be 0
-        { tap_csv_preimage::step8_data_input_part_if_anyonecanpay::step1_outpoint::Step2IndexGadget::from_constant(0) }
-        OP_CAT3
-
-        // get a hint: previous tx's amount
-        OP_HINT
-        OP_SIZE 8 OP_EQUALVERIFY
-        OP_DUP OP_TOALTSTACK
-        OP_CAT2
-
-        // add the script pub key
-        2 OP_PICK
-        OP_PUSHBYTES_1 OP_PUSHBYTES_34 OP_SWAP
-        OP_CAT3
-
-        // require the input sequence number be 0xffffffff
-        { tap_csv_preimage::step8_data_input_part_if_anyonecanpay::Step4SequenceGadget::from_constant(&Sequence::default()) }
-        OP_CAT2
-
-        OP_FROMALTSTACK OP_SWAP
-        OP_FROMALTSTACK OP_SWAP
-
-        // current stack body: this scriptpubkey, previous counter, previous tx's amount, previous tx's txid, |1-11|
-
-        // get a hint: tap leaf hash
-        OP_HINT
-        OP_SIZE 32 OP_EQUALVERIFY
-
-        { tap_csv_preimage::step12_ext::Step2KeyVersionGadget::from_constant(0) }
-        { tap_csv_preimage::step12_ext::Step3CodeSepPosGadget::no_code_sep_executed() }
-        OP_CAT4
-
-        // current stack body: this scriptpubkey, previous counter, previous tx's amount, previous tx's txid, this script's checksighash preimage
-        { TaggedHashGadget::from_provided(&HashTag::TapSighash) }
-
-        { secp256k1_generator.clone() }
-        OP_DUP OP_TOALTSTACK
-        OP_DUP OP_TOALTSTACK
-
-        OP_DUP OP_ROT OP_CAT3
-
-        { TaggedHashGadget::from_provided(&HashTag::BIP340Challenge) }
-
-        // get a hint: the sha256 without the last byte
-        OP_HINT
-        OP_SIZE 31 OP_EQUALVERIFY
-
-        OP_DUP { 1 } OP_CAT
-        OP_ROT OP_EQUALVERIFY
-
-        // current stack body:
-        //   this scriptpubkey, previous counter, previous tx's amount, previous tx's txid,
-        //   prefix
-
-        OP_FROMALTSTACK OP_SWAP
-
-        OP_PUSHBYTES_2 OP_PUSHBYTES_2 OP_RIGHT
-        OP_CAT3
-
-        OP_FROMALTSTACK
-        OP_CHECKSIGVERIFY
-
-        // current stack body:
-        //   this scriptpubkey, previous counter, previous tx's amount, previous tx's txid
-
-        { tx::Step1VersionGadget::from_constant(&Version::ONE) }
-
-        // get a hint: first input's outpoint
-        OP_HINT
-        OP_SIZE 36 OP_EQUALVERIFY
-
-        // get a hint: second input's outpoint
-        OP_HINT
-        OP_SIZE 0 OP_EQUAL OP_TOALTSTACK
-        OP_SIZE 36 OP_EQUAL OP_FROMALTSTACK OP_BOOLOR OP_VERIFY
-
-        OP_SIZE 0 OP_EQUAL
-        OP_IF
-            OP_DROP
-            OP_PUSHBYTES_5 OP_PUSHBYTES_0 OP_INVALIDOPCODE OP_INVALIDOPCODE OP_INVALIDOPCODE OP_INVALIDOPCODE
-            OP_CAT
-            { tx::Step2InCounterGadget::from_constant(1) }
-            OP_SWAP OP_CAT
-        OP_ELSE
-            OP_TOALTSTACK
-            OP_PUSHBYTES_5 OP_PUSHBYTES_0 OP_INVALIDOPCODE OP_INVALIDOPCODE OP_INVALIDOPCODE OP_INVALIDOPCODE
-            OP_DUP
-            OP_FROMALTSTACK OP_SWAP
-            OP_CAT4
-            { tx::Step2InCounterGadget::from_constant(2) }
-            OP_SWAP OP_CAT
-        OP_ENDIF
-        OP_CAT2
-
-        { tx::Step4OutCounterGadget::from_constant(2) }
-        OP_CAT2
-
-        // current stack body:
-        //   this scriptpubkey, previous counter, previous tx's amount, previous tx's txid
-        //   txid preimage (1-4)
-
-        // get the previous amount
-        2 OP_ROLL
-        OP_CAT2
-
-        // get the script pub key
-        3 OP_ROLL
-        OP_PUSHBYTES_1 OP_PUSHBYTES_34 OP_SWAP
-        OP_CAT3
-
-        { tx::step5_output::Step1AmountGadget::from_constant(&Amount::from_sat(DUST_AMOUNT)) }
-        OP_CAT2
-
-        // push the script hash header
-        OP_PUSHBYTES_2 OP_RETURN OP_PUSHBYTES_8
-
-        3 OP_ROLL
-
-        // extend the actual counter to 4 bytes
-        { CppInt32Gadget::from_positive_bitcoin_integer() }
-
-        // get a hint: the randomizer for previous transaction (4 bytes)
-        OP_HINT
-        OP_SIZE 4 OP_EQUALVERIFY
-        OP_CAT3
-
-        OP_SHA256
-
-        OP_PUSHBYTES_3 OP_PUSHBYTES_34 OP_PUSHBYTES_0 OP_PUSHBYTES_32
-        OP_SWAP OP_CAT3
-
-        { tx::Step6LockTimeGadget::from_constant_absolute(&LockTime::ZERO) }
-        OP_CAT2
-
-        OP_SHA256
-        OP_SHA256
-
-        OP_EQUALVERIFY
-
-        OP_FROMALTSTACK OP_FROMALTSTACK
+        covenants
         OP_1SUB OP_EQUAL
     }
 }
@@ -524,7 +264,7 @@ pub fn get_script() -> Script {
 #[cfg(test)]
 mod test {
     use crate::counter::{
-        get_script, get_script_pub_key_and_control_block, get_tx, CovenantHints, DUST_AMOUNT,
+        get_script, get_script_pub_key_and_control_block, get_tx,
     };
     use crate::treepp::*;
     use bitcoin::absolute::LockTime;
@@ -532,8 +272,8 @@ mod test {
     use bitcoin::opcodes::all::{OP_PUSHBYTES_8, OP_RETURN};
     use bitcoin::transaction::Version;
     use bitcoin::{
-        Amount, OutPoint, ScriptBuf, Sequence, Transaction, TxIn, TxOut, Txid, WScriptHash,
-        Witness, WitnessProgram,
+        Amount, OutPoint, ScriptBuf, Sequence, Transaction, Txid, TxIn, TxOut, Witness,
+        WitnessProgram, WScriptHash,
     };
     use bitcoin_scriptexec::{Exec, ExecCtx, Options};
     use bitcoin_simulator::database::Database;
@@ -542,6 +282,7 @@ mod test {
     use rand_chacha::ChaCha20Rng;
     use std::cell::RefCell;
     use std::rc::Rc;
+    use crate::common::{CovenantHints, DUST_AMOUNT};
 
     #[test]
     fn test_script_execution() {
