+++ src/lib.rs
@@ -7,6 +7,8 @@ pub static SECP256K1_GENERATOR: Lazy<Vec<u8>> = Lazy::new(|| {
 
 pub mod counter;
 
+pub mod common;
+
 pub(crate) mod treepp {
     pub use bitcoin_script::{define_pushable, script};
